+++ hit.go
@@ -25,10 +25,6 @@ type HitOptions struct {
 	// Client is the database client required to look up sessions.
 	Client Store
 
-	// Timezone is the the timezone used to create the fingerprint.
-	// It will be set to UTC by default.
-	Timezone *time.Location
-
 	// ClientID is optionally saved with a hit to split the data between multiple clients.
 	ClientID int64
 
@@ -79,17 +75,13 @@ func HitFromRequest(r *http.Request, salt string, options *HitOptions) Hit {
 		options = &HitOptions{}
 	}
 
-	if options.Timezone == nil {
-		options.Timezone = time.UTC
-	}
-
 	if options.SessionMaxAge.Seconds() == 0 {
 		options.SessionMaxAge = defaultSessionMaxAge
 	}
 
 	// shorten strings if required and parse User-Agent to extract more data (OS, Browser)
 	getRequestURI(r, options)
-	fingerprint := Fingerprint(r, salt, options.Timezone)
+	fingerprint := Fingerprint(r, salt)
 	userAgent := r.UserAgent()
 	path := shortenString(options.Path, 2000)
 	requestURL := shortenString(options.URL, 2000)